+++ docs/CHANGELOG.md
@@ -21,6 +21,7 @@ Version 1.2.0 (Willow): Not yet released
 - Add `DesiredHandicap` setting to `alpine_settings.ini`
 - Add `cl_autoswitchfirewait` command to suppress autoswitch during and immediately after weapon fire
 - Add `AutoswitchFireWait` setting to `alpine_settings.ini`
+- Maintain third person camera view between level loads after `camera3` command entered
 
 ### Bug fixes
 [@GooberRF](https://github.com/GooberRF)